syntax: Properly highlight Nix builtin primops.
Builtins listed in the nixFunction keywords so far didn't recognize the
builtins.* namespace and also weren't complete.

This now matches on all the builtins from the manual of Nix 1.10 and
also takes care of the builtins.* namespace.

With one exception: I didn't include isNull, because it's deprecated.

Note that I've added it to the end of the syntax file to ensure that we
don't clash with any earlier matches.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>