Add test.nix file to make testing easier with Nix.
This allows to just run "nix-build test.nix" and we get a Vim version
with isolated dependencies and just Vader and nothing more.

I've used this mainly because on my setup I'm patching in Vim plugins
into the default vimrc itself rather than having a dotfile or a
system-wide file.

So this should avoid any clashes with existing plugins or the stock Vim
initialization file.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>