syntax: Merge non-namespaced+namespaced builtins.
Just stumbled over this in a Nix expression file at:

https://github.com/NixOS/nixpkgs/blob/b843e7c50a3e9793625c4ed5af890a58a9620074/nixos/modules/installer/cd-dvd/iso-image.nix#L35

After confirming this with nix-repl, it's indeed the case that the
builtins.* namespace also contains the non-namespaced primops.

The reason why I didn't add nixSimpleBuiltin to the contains= attribute
for nixBuiltin is that we want to have the same highlighting as for
nixNamespacedBuiltin.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>