syntax: Use space/comment eating regex everywhere.
While I was very reluctant to do this while working on 75e965c, it turns
out that while using it in practice while writing and reading Nix files
it really IS needed, because we have all sorts of commenting/whitespace
everywhere.

I did also insert a few more complications in the tests to make sure we
really get as much of the corner-cases we ever can.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>