syntax: Fix matching < and > operators
So far a search path reference has been a syntax region, which is a very
lax match for the strict rules from the actual Nix expression parser and
if there is a single < operator, the syntax highlighting for the rest of
the file is going to be messed up.

Let's turn it into a "syn match" and highlight via the subgroups.

Of course, this also matches the operators we actually want to match,
which are <, >, <= and >=.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>