Fix indent of "with ..;" in conjunction with "let"
It's not very clear to me what 7452e0808a12cff7883849eaa058777f5754f77e
was trying to fix *exactly* (it also doesn't include tests and the
commit message is just "align let/attribute bindings"), but it
introduced a regression for me because I write Nix expressions like this
on a daily basis:

  with import <nixpkgs> {};

  let
    ...
  in ...

Since that commit the expression is indented like this:

  with import <nixpkgs> {};

    let
      ...
      in ...

Here the indenter is looking on whether the last non-blank line ends
with a semicolon (which is the case here) and then searches a pair of
s:binding_open and s:binding_close (which is also the case here) and
indents the current line with the indentation of s:binding_open plus one
shiftwidth.

The problem however is that when the indenter is called after the let
has been written it recognizes the *current* line as a valid pair and
thus indents it with the *current* indentation plus shiftwidth.

I've fixed this by adding an additional check to not do this whenever a
s:binding_open is in the current line. However, I'm not quite sure
whether this is really sufficient, so it probably would be best to
overhaul this whole logic in the long term.

Signed-off-by: aszlig <aszlig@nix.build>