Highlight string escape sequence errors
Before a2a9e1187b4abdb9d78a6f43c159c9ff14c54167, the check for the
escape characters was quite strict and didn't match escape sequences
where Nix just strips off the backslash (or ''\ for multiline strings).

The strict match is useful if you want to see a change in highlighting
so that you'll notice whenever you use a character that either is
mistyped or the whole escape sequence is unnecessary.

Now we have the old check in again but with a second group
(nixInvalidSimpleStringEscape and nixInvalidStringEscape respectively),
which should now highlight this with the Error highlighting group, so
users will notice immediately.

Signed-off-by: aszlig <aszlig@nix.build>