test/multiline-string-escape: Fix SyntaxOf pattern
When double quotes are used, backslashes need to be escaped twice, so
the pattern for "''\\" would result in the regex /''\/ which is still
valid, but a single backslash in regexes is also used as an escape
character. So just to make sure this is correctly escaped I added
another backslash.

In essence, this doesn't have any changes in functionality, but is just
a small nitpick ;-)

Signed-off-by: aszlig <aszlig@nix.build>