syntax: Fix regex for nixAttribute match
Using dashes before the end of a character collection is quite dangerous
because they represent character ranges.

This is the case for the regex of the nixAttribute group which has been
changed in a61495a762feacc00f24cab4392b09cc3250decf, the '-. now is a
character range and thus accepts any of the following characters:

  * 39 0x27 '
  * 40 0x28 (
  * 41 0x29 )
  * 42 0x2A *
  * 43 0x2B +
  * 44 0x2C ,
  * 45 0x2D -
  * 46 0x2E .

To fix this, I just moved the dash to the end of the character
collection so now only "'", "." and "-" are matched.

Signed-off-by: aszlig <aszlig@nix.build>