Fix highlighting of function-args vs URIs
The following expression is actually a string that needs to be matched
by `nixURI`:

  > a:b

However, this is matched as a function since `nixSimpleFunctionArgument`
was only unmatched if it's followed by a `/`. This change uses the `\@=`
operator[1] to make sure that a function-argument is only detected if
the colon is followed by a newline or a space.

This is currently helpful for e.g. `nix-flakes` where you can define
inputs like this[2]:

  > { input.nixpkgs.url = github:NixOS/nixpkgs/nixos-20.03; }

[1] http://vimdoc.sourceforge.net/htmldoc/pattern.html#/\@=
[2] https://www.tweag.io/blog/2020-05-25-flakes/