nixExpr: highlight `args@{ ... }` in an assignment
To reproduce the original issue:

The expression

    let
      foo = args@{ ... }: args;
    in foo

doesn't get highlighted, whereas

    let
      foo = { ... }@args: args;
    in foo

does.

The issue is that `nixArgOperator` should be matchable right in the
cluster `nixExpr` which is used when an assignment in an attr-set or a
`let-in` expression happens. It's ensured that only correct syntax
syntax gets highlighted since `nixArgOperator` requires
`nixFunctionArgument` after that.