Switch from travis to GHA
In fact, the tests are currently failing on `master` and we didn't
notice because travis is dead and thus the CI isn't running anymore.

Added a simple GHA pipeline to run this.

Also altered the `default.nix` to not use `fetchGit` anymore: for such a
small package this doesn't bring any difference and GHA does shallow
clones by default, so we can't use it as-is anyways.