Fix nixInherit matcher
The test for `inherit (foo) bar`-expressions is currently failing with

    ( 8/37) [EXECUTE] (X) 'nixInheritAttributeScope' should be equal to 'nixParen'

This is because the opening bracket is also matched by
`nixExpr->nixParen`. The workaround is to perform the lookahead in
`nixInheritAttributeScope` and then hand over to a new region called
`nixInheritAttributeSubExpr` which sets the match start to one char
after the opening bracket to avoid the double-match.

Finally, only do a lookahead to `)` in `nixInheritAttributeSubExpr` (and
thus make sure the region is closed to not get a match of `nixParen`
here) and let `nixInheritAttributeScope` close the bracket.S

The code itself worked fine, but the matchings were off, so if you'd
highlight e.g. `nixInheritAttributeScope` in your `.vimrc`, the breakage
would be noticeable.

Finally, we need to prepend the runtime-path now because `nix` is now
part of `vim/vim`[1].

[1] https://github.com/vim/vim/commit/86b4816766d976a7ecd4403eca1f8bf6b4105800#diff-5202ae314163489b5d9872596ccd1bcd9ec8215c64722447dc7ba3b6cbceadd5