Fix nixInherit matcher
The test for `inherit (foo) bar`-expressions is currently failing with

    ( 8/37) [EXECUTE] (X) 'nixInheritAttributeScope' should be equal to 'nixParen'

This is because the opening bracket is also matched by
`nixExpr->nixParen`. The workaround is to perform the lookahead in
`nixInheritAttributeScope` and then hand over to a new region called
`nixInheritAttributeSubExpr` which sets the match start to one char
after the opening bracket to avoid the double-match.

Finally, only do a lookahead to `)` in `nixInheritAttributeSubExpr` (and
thus make sure the region is closed to not get a match of `nixParen`
here) and let `nixInheritAttributeScope` close the bracket.